remove __attribute__((weak)) from headers
Weak attributes should only go on the symbols you _actually_ want
to mark as weak, i.e. those in the C sources.

By including this in the headers, you actually cause _every_ definition
of these symbols to be weak, in which case the linker just picks one.

This was likely working without anyone noticing because people generally
write the link rules `ld mystuff.o libnrf.a`, which means the linker
sees your symbol first and happens to choose that one, but that's just
luck really.